Fixed location for Availability Set
The Availability Set was being set by the Resource Group location. This meant that if you specified one region in the azureDeploy.parameters.json file but the Resource Group location was elsewhere, the deployment would fail due to the Availability Set being in a different location. The Availability Set now looks up the location parameter from the azureDeploy.parameters.json file and follows the location of the other resources.